+++ src/CpuUsageCalculator/CpuUsageCalculator.cpp
@@ -5,11 +5,11 @@
 
 using namespace get_system_usage_linux;
 
-int CpuUsageCalculator::GetPercentageCpu()
+float CpuUsageCalculator::GetPercentageCpu()
 {
     CPU_stats t1 = read_cpu_data();
     std::this_thread::sleep_for(std::chrono::milliseconds(1000));
     CPU_stats t2 = read_cpu_data();
 
-    return 100 * get_cpu_usage(t1, t2);
+    return 100.f * get_cpu_usage(t1, t2);
 }
